Refactor Observability utilities into a separate directory under src/
