+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/MetricsGraphs.js
@@ -19,7 +19,7 @@
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import CardHeader from "@material-ui/core/CardHeader";
-import Constants from "../common/constants";
+import Constants from "../../utils/constants";
 import Grid from "@material-ui/core/Grid";
 import InfoIcon from "@material-ui/icons/InfoOutlined";
 import React from "react";
