+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/cell/Details.js
@@ -17,9 +17,9 @@
 import CellDependencyView from "./CellDependencyView";
 import ColorGenerator from "../../common/color/colorGenerator";
 import HealthIndicator from "../../common/HealthIndicator";
-import HttpUtils from "../../common/utils/httpUtils";
-import NotificationUtils from "../../common/utils/notificationUtils";
-import QueryUtils from "../../common/utils/queryUtils";
+import HttpUtils from "../../../utils/common/httpUtils";
+import NotificationUtils from "../../../utils/common/notificationUtils";
+import QueryUtils from "../../../utils/common/queryUtils";
 import React from "react";
 import StateHolder from "../../common/state/stateHolder";
 import Table from "@material-ui/core/Table";
