+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/microservice/Details.js
@@ -15,10 +15,10 @@
  */
 
 import HealthIndicator from "../../common/HealthIndicator";
-import HttpUtils from "../../common/utils/httpUtils";
+import HttpUtils from "../../../utils/common/httpUtils";
 import {Link} from "react-router-dom";
-import NotificationUtils from "../../common/utils/notificationUtils";
-import QueryUtils from "../../common/utils/queryUtils";
+import NotificationUtils from "../../../utils/common/notificationUtils";
+import QueryUtils from "../../../utils/common/queryUtils";
 import React from "react";
 import ServiceDependencyView from "./ServiceDependencyView";
 import StateHolder from "../../common/state/stateHolder";
