+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/microservice/index.js
@@ -17,7 +17,7 @@
 import Button from "@material-ui/core/Button/Button";
 import Details from "./Details";
 import Grey from "@material-ui/core/colors/grey";
-import HttpUtils from "../../common/utils/httpUtils";
+import HttpUtils from "../../../utils/common/httpUtils";
 import K8sObjects from "./K8sObjects";
 import {Link} from "react-router-dom";
 import Metrics from "./Metrics";
