+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/common/state/stateHolder.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import AuthUtils from "../utils/authUtils";
+import AuthUtils from "../../../utils/common/authUtils";
 
 /**
  * Configuration holder.
