+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/overview/SidePanelContent.js
@@ -27,7 +27,7 @@ import ExpansionPanel from "@material-ui/core/ExpansionPanel";
 import ExpansionPanelDetails from "@material-ui/core/ExpansionPanelDetails";
 import ExpansionPanelSummary from "@material-ui/core/ExpansionPanelSummary";
 import Grey from "@material-ui/core/colors/grey";
-import HttpUtils from "../common/utils/httpUtils";
+import HttpUtils from "../../utils/common/httpUtils";
 import IconButton from "@material-ui/core/IconButton";
 import {Link} from "react-router-dom";
 import MUIDataTable from "mui-datatables";
