+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/overview/index.js
@@ -20,20 +20,20 @@ import ArrowRightAltSharp from "@material-ui/icons/ArrowRightAltSharp";
 import Button from "@material-ui/core/Button";
 import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-import Constants from "../common/constants";
+import Constants from "../../utils/constants";
 import DependencyGraph from "../common/DependencyGraph";
 import Divider from "@material-ui/core/Divider";
 import Drawer from "@material-ui/core/Drawer";
 import Error from "@material-ui/icons/Error";
 import Fade from "@material-ui/core/Fade";
 import Grey from "@material-ui/core/colors/grey";
-import HttpUtils from "../common/utils/httpUtils";
+import HttpUtils from "../../utils/common/httpUtils";
 import IconButton from "@material-ui/core/IconButton";
 import MoreIcon from "@material-ui/icons/MoreHoriz";
-import NotificationUtils from "../common/utils/notificationUtils";
+import NotificationUtils from "../../utils/common/notificationUtils";
 import Paper from "@material-ui/core/Paper";
 import Popper from "@material-ui/core/Popper";
-import QueryUtils from "../common/utils/queryUtils";
+import QueryUtils from "../../utils/common/queryUtils";
 import React from "react";
 import SidePanelContent from "./SidePanelContent";
 import StateHolder from "../common/state/stateHolder";
