+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/Search.js
@@ -18,17 +18,17 @@ import Button from "@material-ui/core/Button";
 import ChipInput from "material-ui-chip-input";
 import FormControl from "@material-ui/core/FormControl/FormControl";
 import Grid from "@material-ui/core/Grid/Grid";
-import HttpUtils from "../common/utils/httpUtils";
+import HttpUtils from "../../utils/common/httpUtils";
 import InputAdornment from "@material-ui/core/InputAdornment/InputAdornment";
 import InputLabel from "@material-ui/core/InputLabel/InputLabel";
 import MenuItem from "@material-ui/core/MenuItem/MenuItem";
-import NotificationUtils from "../common/utils/notificationUtils";
+import NotificationUtils from "../../utils/common/notificationUtils";
 import Paper from "@material-ui/core/Paper/Paper";
-import QueryUtils from "../common/utils/queryUtils";
+import QueryUtils from "../../utils/common/queryUtils";
 import React from "react";
 import SearchResult from "./SearchResult";
 import Select from "@material-ui/core/Select/Select";
-import Span from "./utils/span";
+import Span from "../../utils/tracing/span";
 import TextField from "@material-ui/core/TextField/TextField";
 import TopToolbar from "../common/toptoolbar";
 import Typography from "@material-ui/core/Typography/Typography";
