+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/DependencyDiagram.js
@@ -17,7 +17,7 @@
 import "./DependencyDiagram.css";
 import ArrowRightAltSharp from "@material-ui/icons/ArrowRightAltSharp";
 import Button from "@material-ui/core/Button";
-import Constants from "../../common/constants";
+import Constants from "../../../utils/constants";
 import Error from "@material-ui/icons/Error";
 import Fade from "@material-ui/core/Fade";
 import FiberManualRecord from "@material-ui/icons/FiberManualRecord";
@@ -26,8 +26,8 @@ import Paper from "@material-ui/core/Paper";
 import Popper from "@material-ui/core/Popper";
 import PropTypes from "prop-types";
 import React from "react";
-import Span from "../utils/span";
-import TracingUtils from "../utils/tracingUtils";
+import Span from "../../../utils/tracing/span";
+import TracingUtils from "../../../utils/tracing/tracingUtils";
 import Typography from "@material-ui/core/Typography";
 import {withStyles} from "@material-ui/core";
 import withColor, {ColorGenerator} from "../../common/color";
