+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/index.js
@@ -17,18 +17,18 @@
 import DependencyDiagram from "./DependencyDiagram";
 import ErrorBoundary from "../../common/error/ErrorBoundary";
 import Grey from "@material-ui/core/colors/grey";
-import HttpUtils from "../../common/utils/httpUtils";
+import HttpUtils from "../../../utils/common/httpUtils";
 import NotFound from "../../common/error/NotFound";
-import NotificationUtils from "../../common/utils/notificationUtils";
+import NotificationUtils from "../../../utils/common/notificationUtils";
 import Paper from "@material-ui/core/Paper/Paper";
 import React from "react";
 import SequenceDiagram from "./sequenceDiagram/SequenceDiagram";
-import Span from "../utils/span";
+import Span from "../../../utils/tracing/span";
 import Tab from "@material-ui/core/Tab";
 import Tabs from "@material-ui/core/Tabs";
 import Timeline from "./timeline";
 import TopToolbar from "../../common/toptoolbar";
-import TracingUtils from "../utils/tracingUtils";
+import TracingUtils from "../../../utils/tracing/tracingUtils";
 import UnknownError from "../../common/error/UnknownError";
 import withStyles from "@material-ui/core/styles/withStyles";
 import withGlobalState, {StateHolder} from "../../common/state";
