+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/timeline/TimelineView.js
@@ -20,16 +20,16 @@ import "vis/dist/vis-timeline-graph2d.min.css";
 import "./TimelineView.css";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
-import Constants from "../../../common/constants";
+import Constants from "../../../../utils/constants";
 import PropTypes from "prop-types";
 import React from "react";
 import ReactDOM from "react-dom";
-import Span from "../../utils/span";
+import Span from "../../../../utils/tracing/span";
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
 import TableRow from "@material-ui/core/TableRow";
-import TracingUtils from "../../utils/tracingUtils";
+import TracingUtils from "../../../../utils/tracing/tracingUtils";
 import Typography from "@material-ui/core/Typography";
 import interact from "interactjs";
 import vis from "vis";
