+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/common/authUtils.js
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import {StateHolder} from "../state";
+import {StateHolder} from "../../components/common/state";
 
 /**
  * Authentication/Authorization related utilities.
