+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/common/authUtils.test.js
@@ -17,7 +17,7 @@
  */
 
 import AuthUtils from "./authUtils";
-import {StateHolder} from "../state";
+import {StateHolder} from "../../components/common/state";
 
 describe("AuthUtils", () => {
     afterEach(() => {
