+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/common/httpUtils.js
@@ -15,7 +15,7 @@
  */
 
 import AuthUtils from "./authUtils";
-import {StateHolder} from "../state";
+import {StateHolder} from "../../components/common/state";
 import axios from "axios";
 
 class HttpUtils {
