+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/common/notificationUtils.test.js
@@ -15,7 +15,7 @@
  */
 
 import NotificationUtils from "./notificationUtils";
-import StateHolder from "../state/stateHolder";
+import StateHolder from "../../components/common/state/stateHolder";
 
 describe("NotificationUtils", () => {
     let state;
