+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/constants.js
