+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/utils/tracing/tracingUtils.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import Constants from "../../common/constants";
+import Constants from "../constants";
 
 /**
  * Utilities used for processing Constants related data.
